Cleanup arangodb database interface
